refactor: ui tests - letters
#2037
